+++ package.json
@@ -7,7 +7,7 @@
     "develop": "gulp develop",
     "generate-changelog": "changelog-maker --group",
     "test": "gulp test",
-    "demos": "webpack-dev-server"
+    "demo": "webpack-dev-server"
   },
   "repository": {
     "type": "git",
@@ -19,17 +19,13 @@
     "npm": "3.10.8"
   },
   "dependencies": {
-    "promise-polyfill": "6.0.2",
-    "whatwg-fetch": "2.0.3"
+    "@webcomponents/webcomponentsjs": "1.0.1"
   },
   "devDependencies": {
     "chai": "3.5.0",
     "babel-cli": "6.24.1",
     "babel-core": "6.24.1",
     "babel-loader": "7.0.0",
-    "babel-plugin-transform-object-assign": "6.8.0",
-    "babel-plugin-transform-runtime": "6.23.0",
-    "babel-polyfill": "6.23.0",
     "babel-preset-es2015": "6.6.0",
     "eslint": "3.19.0",
     "eslint-loader": "1.7.1",
