+++ src/next/index.js
@@ -1,5 +1,4 @@
 require('./boot.js');
 require('./utils.js');
-require('./inject-dom.js');
 require('./inject-web-components.js');
 require('./inject-bundles.js');
